requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
tenacity==8.5.0
trafilatura==1.7.0
readability-lxml==0.8.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
openai==1.45.0
numpy==1.26.4
pytest==8.3.2
